Patch SupraLib for a 3072-element 4D TEE matrix-array probe
Small, targeted patches needed to integrate SUPRA as the beamforming
backend for TEERegCline's TEE 4D matrix probe bridge (see
Digisurge/TEERegCline, src/supra_bridge):

- Beamformer/RxBeamformerCuda.cu: the 3D delay-and-sum kernel hardcoded
  its shared-memory element cache to 1024 entries, silently corrupting
  results for any array with more elements. Raised to 3072 (named
  kTeeMaxRxElements) to cover the probe's element count; well under the
  48KB per-block shared-memory budget on every targeted architecture.
- Beamformer/ScanConverter.h: added a getBoundingBoxMin() getter so two
  independently scan-converted volumes (full B-mode sweep vs. a small
  Doppler ROI) can be aligned onto one combined voxel grid.
- SupraManager.h: addNodeConstruct<T>() took its constructor argument
  pack by value and forwarded it as lvalues, which cannot bind a
  move-only argument (e.g. std::unique_ptr). Changed to perfect
  forwarding; behavior for existing copyable-argument callers is
  unchanged.
- SupraLib/CMakeLists.txt: two build-system fixes -- (1) legacy
  CUDA_ADD_LIBRARY() does not see target_compile_definitions() added
  after the fact, so HAVE_CUDA was silently unset while compiling .cu
  files; added add_definitions() before the library is declared.
  (2) TBB's headers pull in <immintrin.h>, which on newer glibc/gcc
  unconditionally drags in AMX intrinsic headers that nvcc's front-end
  cannot parse; pre-defined those headers' include guards as a no-op
  (AMX is unused by any .cu file here).
- SupraLib/NoiseCuda.cu: one NPP API out-param changed from int* to
  size_t* between the 2016-era API this was written against and CUDA 12.

Each change is marked "[TEE patch]" inline with the same rationale.

+++ src/SupraLib/NoiseCuda.cu
@@ -135,7 +135,9 @@ namespace supra
 
 		auto meanStdDevBefore = make_shared<Container<double> >(LocationGpu, in->getStream(), 2);
 		auto meanStdDevAfter = make_shared<Container<double> >(LocationGpu, in->getStream(), 2);
-		int scratchSize = 0;
+		// [TEE patch] CUDA 12's NPP headers changed this out-param from int* to
+		// size_t* (it was int* in the 2016-era NPP API this code was written for).
+		size_t scratchSize = 0;
 		nppSafeCall(nppiMeanStdDevGetBufferHostSize_32f_C1R({ iHeight, iWidth*iDepth }, &scratchSize));
 		auto meanStdDevScratch = make_shared<Container<uint8_t> >(LocationGpu, in->getStream(), scratchSize);
